+++ ElvUI_MerathilisUI/Core/Changelog/Previous/6.50.lua
@@ -1,0 +1,15 @@
+local MER = unpack(ElvUI_MerathilisUI)
+
+MER.Changelog[650] = {
+	RELEASE_DATE = "10.01.2025",
+	FIXES = {
+		"[World Map]: Try to avoid an error in Navigation Bar",
+		"[Style]: My style should now also apply on UnitFrames again",
+	},
+	NEW = {
+		"[Socket Frame]: Adds a selection Frame on the Socketing Frame for the new singing Gems",
+	},
+	IMPROVEMENTS = {
+		"[Misc]: Remove some useless code in Talking Head",
+	},
+}