+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changes ###
 
-  * [World Map]: Updated the Navigation Bar look on the top
-  * [World Map]: Updated the Map Reveal Data
-  * [AutoButtons]: Add more reward boxes
+  * [World Map]: Try to avoid an error in Navigation Bar
+  * [Style]: My style should now also apply on UnitFrames again
+  * [Socket Frame]: Adds a selection Frame on the Socketing Frame for the new singing Gems
+  * [Misc]: Remove some useless code in Talking Head